requirements.txt
# requirements.txt
Flask==2.1.0
qrcode==7.3
Pillow==9.0.0